requirements.txt
# Core system dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.4.0
fastapi>=0.100.0
uvicorn>=0.23.0
rich>=13.0.0
typer>=0.9.0
networkx>=3.1
plotly>=5.15.0
streamlit>=1.28.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Sparse search dependencies
rank-bm25>=0.2.2
nltk>=3.8

# Additional development tools
jupyterlab>=4.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Code quality
pre-commit>=3.3.0
bandit>=1.7.0
safety>=2.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Performance profiling
memory-profiler>=0.61.0
line-profiler>=4.0.0

# Testing coverage
codecov>=2.1.0

matplotlib
seaborn




